Fix SQL Server CAST() tokens and extract cast-type mapping from PlatformCapabilities
getSupportedCastTypes() had no sqlsrv branch, so it silently fell into
the MySQL/MariaDB default and emitted CAST(x AS SIGNED)/CAST(x AS
DOUBLE) - both invalid T-SQL, since SIGNED/DOUBLE are MySQL-only CAST
keywords. Add an explicit sqlsrv branch: INT/FLOAT (the real T-SQL
names), NVARCHAR(MAX) for string (bare CHAR silently truncates to 30
chars in T-SQL), and BIT for bool (SQL Server has a real boolean-ish
CAST target, unlike MySQL/MariaDB which have none at all).

Also extract the cast-type lookup table itself into a new
CastTypeMapper, alongside DDLTypeMapper/SqlIdentifierQuoter, since it's
the same kind of thing (an abstract-type-to-SQL-token map) that had
already been pulled out of PlatformCapabilities for temp-table DDL and
identifier quoting - PlatformCapabilities should only report booleans/
tokens, not carry per-type lookup tables. BuildSqlFromAst and
SemanticAnalyzer build a CastTypeMapper from their existing
PlatformCapabilitiesInterface instead of calling the removed method
directly.

Condense DDLTypeMapper's comments to keep only the load-bearing
rationale (SQL Server's '#'-prefix and TIMESTAMP/rowversion traps,
Postgres's missing 1-byte int type, SQLite's affinity model).
